Extract ClickStack scalar enums
